test/requirements.txt
Werkzeug<1.0.0
sqlalchemy==1.3.23 # hotfix for initdb regression in sqlalchemy 1.4
apache-airflow==1.10.7
botocore==1.16.4
boto3==1.13.4
moto==1.3.14
pytest-pythonpath==0.7.3
pypandoc==1.5
requests-mock==1.8.0
pysolar==0.8
